#233 컴포넌트화
