+++ pubspec.yaml
@@ -32,7 +32,7 @@ dependencies:
   path_provider:
   random_string: ^2.0.1
   animated_text_kit: ^2.2.0
-  provider: ^3.0.0+1
+  provider: ^4.3.2+3
   charts_flutter: ^0.9.0
   flutter_native_splash: ^0.1.9
   uuid: ^2.2.0
@@ -42,7 +42,7 @@ dependencies:
   numberpicker: ^1.2.0
   #  flutter_rounded_date_picker: 1.0.0
   fluttertoast: ^7.0.1
-  flutter_svg: ^0.18.0
+  flutter_svg: ^0.19.0
   firebase_messaging: ^6.0.1
   kakao_flutter_sdk: ^0.5.2
 
@@ -51,6 +51,8 @@ dependencies:
   cupertino_icons: ^0.1.3
   flutter_sparkline: ^0.1.0
   collection: ^1.14.13
+  intl: ^0.15.7
+  calendar_widget: ^0.1.0
 
 dev_dependencies:
   test:
